Implement Management Semester, Subject, Combo, Reading data from .CSV file. Fixing Blind Prompt of RAG  AI Chat

+++ src/main/java/AiStudyHub/BE/entity/Ranking.java
@@ -22,6 +22,6 @@ public class Ranking {
     Integer minScore;
     Integer maxScore;
     @Builder.Default
-    Integer storageBonus = 0;
+    Long storageBonus = 0L;
     String displayPriority;
 }
